addressed more review comments
